// experimental/dominion_app/src/types/card-type.ts
export type DCardType =
  | 'Curse'
  | 'Action'
  | 'Treasure'
  | 'Victory'
  | 'Attack'
  | 'Reaction'
  | 'Duration'
  | 'Ruins'
  | 'Prize'
  | 'Looter'
  | 'Shelter'
  | 'Knights'
  | 'Reserve'
  | 'Traveller'
  | 'Castle'
  | 'Gather'
  | 'EventCards'
  | 'LandmarkCards'
  | 'Spirit'
  | 'Night'
  | 'Fate'
  | 'Doom'
  | 'Heirloom'
  | 'Zombie'
  | 'Boon'
  | 'Hex'
  | 'State'
  | 'Artifact'
  | 'Project'